requirements.txt
csv
ctypes
google-auth
google-cloud
math
numpy~=1.20.1
oauthlib
os
pandas~=1.2.3
pandas-datareader
pandas-gbq
plotly~=5.5.0
regex
requests
requests-oauthlib
tkinter
DateTime~=4.3
Pillow~=8.1.2